requirements-lock.txt
aiohappyeyeballs==2.4.4
aiohttp>=3.9.5
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.7.0
apprise==1.9.3
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==24.3.0
Authlib==1.3.1
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cached-property==2.0.1
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
commonmark==0.9.1
coolname==2.2.0
cryptography==44.0.0
dataclasses-json==0.6.7
dateparser==1.2.1
Deprecated==1.2.18
deprecation==2.1.0
dictdiffer==0.9.0
distlib==0.3.9
distro==1.9.0
docker==7.1.0
dotenv==0.9.9
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi>=0.115.0
filelock==3.18.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2025.3.2
gotrue==2.12.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.7.2
grpcio==1.68.1
grpcio-health-checking==1.68.1
grpcio-tools==1.68.1
h11>=0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
humanize==4.12.2
hyperframe==6.1.0
identify==2.6.10
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
ipfshttpclient==0.7.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.46
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.2.10
letta-client==0.1.89
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
md2notion==2.4.1
mdurl==0.1.2
mem0ai==0.1.79
mistletoe==1.4.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.1.0
mypy-extensions==1.0.0
netaddr==1.3.0
networkx==3.4.2
nodeenv==1.9.1
notion==0.0.28
numba==0.61.2
numpy==2.2.1
oauthlib==3.2.2
openai==1.67.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opentelemetry-api==1.31.1
orjson==3.10.14
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
portalocker==2.10.1
postgrest==1.0.1
posthog==3.23.0
pre_commit==4.2.0
prometheus_client==0.21.1
propcache==0.2.1
protobuf==5.29.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.1
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.7.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.13.3
readchar==4.2.1
realtime==2.4.3
referencing==0.36.2
regex==2024.11.6
requests>=2.32.0
requests>=2.32.0
requests>=2.32.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
safetensors==0.5.3
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
sympy==1.14.0
tenacity==9.0.0
text-unidecode==1.3
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers>=4.44.2
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.13.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uuid7==0.1.0
uv==0.6.12
uvicorn>=0.32.0
validators==0.34.0
varint==1.0.2
virtualenv==20.30.0
waitress==3.0.2
weaviate-client==3.26.7
websockets==14.2
Werkzeug==3.1.3
whenever==0.7.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
